trying to update character position

+++ Assets/Scripts/Frameworks/Dtos/CharacterDto.cs
@@ -10,7 +10,7 @@ namespace Frameworks.Dtos
         public EnumCharacterType Type { get; }
         public EnumCharacterDirection Direction { get; }
         public Vector3 Position { get; }
-        public float Speed { get; }
+        public int Speed { get; }
         public string Image { get; }
     }
 
@@ -20,10 +20,10 @@ namespace Frameworks.Dtos
         public EnumCharacterType Type { get; private set; }
         public EnumCharacterDirection Direction { get; private set; }
         public Vector3 Position { get; private set; }
-        public float Speed { get; private set; }
+        public int Speed { get; private set; }
         public string Image { get; private set; }
 
-        private CharacterDto(Guid id, EnumCharacterType type, EnumCharacterDirection direction, Vector3 position, float speed, string image)
+        private CharacterDto(Guid id, EnumCharacterType type, EnumCharacterDirection direction, Vector3 position, int speed, string image)
         {
             this.Id = id;
             this.Type = type;
@@ -33,7 +33,7 @@ namespace Frameworks.Dtos
             this.Image = image;
         }
 
-        public static CharacterDto Create(Guid id, EnumCharacterType type, EnumCharacterDirection direction, Vector3 position, float speed)
+        public static CharacterDto Create(Guid id, EnumCharacterType type, EnumCharacterDirection direction, Vector3 position, int speed)
         {
             string image;
             switch (type)
